requirements.txt
groq>=0.9.0
python-dotenv>=1.1.0
transformers>=4.40.0
peft>=0.10.0
trl>=0.8.0
datasets>=2.19.0
bitsandbytes>=0.43.0
accelerate>=0.29.0
torch>=2.6.0
scipy>=1.13.0
